Strategy Pattern has been applied to monitor the authorization on modifying quizzes, questions, and answer options.

+++ SimpleQuiz.Api/Features/AnswerOptions/DeleteAnswerOption.cs
@@ -1,8 +1,10 @@
 ﻿using Carter;
 using MediatR;
 using SimpleQuiz.Api.Abstractions;
+using SimpleQuiz.Api.Abstractions.Authorizations;
 using SimpleQuiz.Api.Abstractions.Operations;
 using SimpleQuiz.Api.Entities;
+using SimpleQuiz.Api.Services;
 using SimpleQuiz.Api.Shared;
 
 namespace SimpleQuiz.Api.Features.AnswerOptions;
@@ -14,19 +16,33 @@ public static class DeleteAnswerOption
     internal sealed class Handler : ICommandHandler<Command>
     {
         private readonly IRepository<AnswerOption> _answerRepository;
+        private readonly IAuthorizationService<int> _authorizationService;
 
-        public Handler(IRepository<AnswerOption> answerRepository)
+        public Handler(
+            IRepository<AnswerOption> answerRepository,
+            IAuthorizationService<int> authorizationService)
         {
             _answerRepository = answerRepository;
+            _authorizationService = authorizationService;
         }
 
         public async Task<Result> Handle(Command request, CancellationToken cancellationToken)
         {
+
             var answer = await _answerRepository.GetByIdAsync(request.id);
 
             if (answer is null)
             {
                 return Result.Failure(new Error("Invalid answer option id","No answer id has found"));
+            }
+
+            var isAuthorized = await _authorizationService
+                                   .IsUserAuthorizedToModifyAsync
+                                   <QuestionAuthorizationStrategy>(answer.QuestionId);
+
+            if (!isAuthorized)
+            {
+                return Result.Failure(new Error("Authorization.Failure", "User is not authorized to delete answer option on this question."));
             }
 
             await _answerRepository.DeleteAsync(answer);
